Remove bucketIndexInflatedSeries
